Fix Phase 3 bugs and update docs with real-world learnings
- Fix Hibernate schema validation: DECIMAL(10,7) → DOUBLE PRECISION for lat/lng
- Fix native query single-column cast in train run generation
- Fix LazyInitializationException: passengers eager fetch for small collections
- Fix Kafka dual-listener for Docker/host networking (Kafka UI now works)
- Fix passenger status not updating on payment success/failure
- Add error logging to GlobalExceptionHandler catch-all
- Add JOIN FETCH queries to BookingRepository
- Update ARCHITECTURE.md and LEARNINGS.md with Phase 3 bug fixes section

+++ railway-app/src/main/resources/db/migration/V2__create_stations_trains.sql
@@ -5,8 +5,8 @@ CREATE TABLE stations (
     city            VARCHAR(100),
     state           VARCHAR(100),
     zone            VARCHAR(50),
-    latitude        DECIMAL(10, 7),
-    longitude       DECIMAL(10, 7),
+    latitude        DOUBLE PRECISION,
+    longitude       DOUBLE PRECISION,
     is_active       BOOLEAN DEFAULT TRUE,
     created_at      TIMESTAMPTZ NOT NULL DEFAULT NOW()
 );
